ensure invalid settings arnt saved to localstorage
